Added a Person class with Java validation properties, and a webform for it
The class is a standard affair with @Min(x,y) type constraints on
fields.  The webform will redirect to itself if there are errors and
show (ugly) built in java errors (try typing a character on the numeric
field).  For success, a redirect is used to a new url.

A PR was made from this, see Navya's github.